% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_counts.R
\name{adjust_counts}
\alias{adjust_counts}
\title{Use TCRsimilift}
\usage{
adjust_counts(df, sim_method = "HAMMING", cutoff = 0.8)
}
\arguments{
\item{df}{Input dataframe of AIRR formatted immunological data.}

\item{sim_method}{Either HAMMING or BLOSUM.}

\item{cutoff}{Minimum similarity score to consider two sequences neighbours. Between 0 and 1. Default 0.8 .}
}
\value{
Returns dataframe with extra column of adjusted counts based on similarity.
}
\description{
Convenience function to do all TCRsimilift data processing in one call.
The function checks the data to ensure format, prepares the data for processing,
and runs weight_counts to return an updated dataframe with similarity-adjusted
counts.
}
\examples{
results <- adjust_counts(mouse_PBSvTCZ_data_minisubset, sim_method="HAMMING", cutoff = 0.77)

}
