% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosum_similarity.R
\name{blosum_similarity}
\alias{blosum_similarity}
\title{Calculate similarity matrix for input sequences based on BLOSUM distance}
\usage{
blosum_similarity(vj_data)
}
\arguments{
\item{vj_data}{Data subset containing all sequences with same length, same V and J calls.}
}
\value{
Similarity score matrix.
}
\description{
Calculate similarity matrix for input sequences based on BLOSUM distance
}
