% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dist_matrix.R
\name{compute_dist_matrix}
\alias{compute_dist_matrix}
\title{Add column of adjusted counts based on similarity}
\usage{
compute_dist_matrix(data.1, sim_method, cutoff = 0.8)
}
\arguments{
\item{data.1}{Dataframe with counts, AIRR format.}

\item{sim_method}{Either HAMMING or BLOSUM.}

\item{cutoff}{Minimum similarity score to consider two sequences neighbours. Between 0 and 1. Default 0.8 .}
}
\value{
data.1 with added column wrc of adjusted counts.
}
\description{
For cdr3 amino acid length in the sample, and for each existing VJ combination,
a similarity score matrix is calculated for the respective cdr3s. This similarity
matrix is used to generate updated counts via the update_read_counts() function.
}
