% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck.R
\name{datacheck}
\alias{datacheck}
\title{Check if dataset has all necessary columns}
\usage{
datacheck(df)
}
\arguments{
\item{df}{Dataframe of immunological data.}
}
\description{
Check if dataset has all necessary columns
}
