% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{dataprep}
\alias{dataprep}
\title{Data preparation for TCRsimilift}
\usage{
dataprep(df)
}
\arguments{
\item{df}{AIRR formatted dataframe of immunological data.}
}
\value{
A dataframe
}
\description{
Data preparation for TCRsimilift
}
