% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_similarity.R
\name{hamming_similarity}
\alias{hamming_similarity}
\title{Calculate similarity matrix for input sequences based on Hamming distance}
\usage{
hamming_similarity(vj_data)
}
\arguments{
\item{vj_data}{Data subset containing all sequences with same length, same V and J calls.}
}
\value{
Similarity score matrix.
}
\description{
Calculate similarity matrix for input sequences based on Hamming distance
}
