% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_update_data.R
\name{net_update_data}
\alias{net_update_data}
\title{Call compute_dist_matrix for each sample.}
\usage{
net_update_data(data, sim_method = "HAMMING", cutoff = 0.8)
}
\arguments{
\item{data}{Dataframe of AIRR format immunological data.}

\item{sim_method}{Either HAMMING or BLOSUM.}

\item{cutoff}{Minimum similarity score to consider two sequences neighbours. Between 0 and 1. Default 0.8 .}
}
\value{
Returns dataframe with adjusted counts.
}
\description{
Creates a new.data empty dataframe. Then, for each sample, it updates the rows
of the input dataframe with adjusted counts, and appends them to new.data .
In the end, new.data is returned, consisting of all original input rows, but
with an added wrc column for similarity-based adjusted counts.
}
\details{
See \link{TCRsimilift_calculate} for a full example of the DGE workflow.
}
