% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simweightR-package.R
\docType{package}
\name{simweightR-package}
\alias{simweightR}
\alias{simweightR-package}
\title{simweightR: TCR sequencing data preprocessing via similarity-based count adjustment}
\description{
Preprocessing TCRseq data with similarity-based counts weighting for differential abundance analysis.
}
\keyword{internal}
