% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRsimilift-package.R
\docType{package}
\name{TCRsimilift-package}
\alias{TCRsimilift}
\alias{TCRsimilift-package}
\title{TCRsimilift: TCR sequencing data preprocessing via similarity-based count adjustment}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Johannes Textor \email{johannes.textor@ru.nl}

Authors:
\itemize{
  \item Radu Diaconescu \email{r.diaconescu@stud.uni-goettingen.de}
  \item Franka Buytenhuijs
  \item Ankur Ankan
}

}
\keyword{internal}
