% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRsimilift_calculate.R
\name{TCRsimilift_calculate}
\alias{TCRsimilift_calculate}
\title{Use TCRsimilift}
\usage{
TCRsimilift_calculate(
  df,
  sim_method = "HAMMING",
  cutoff = 0.8,
  export_results = FALSE,
  output_directory = "outputs",
  csv_output = FALSE
)
}
\arguments{
\item{df}{Input dataframe of AIRR formatted immunological data.}

\item{sim_method}{Either HAMMING or BLOSUM.}

\item{cutoff}{Minimum similarity score to consider two sequences neighbours. Between 0 and 1. Default 0.8 .}

\item{export_results}{Boolean, whether to automatically run the export_results function.}

\item{output_directory}{Name of output directory.}

\item{csv_output}{TRUE/FALSE. Do you want to export a gzipped csv of the full results? FALSE by default.}
}
\value{
Returns dataframe with extra column of adjusted counts based on similarity.
}
\description{
Convenience function to do all TCRsimilift data processing in one call.
The function checks the data to ensure format, prepares the data for processing,
runs net_update_data to return an updated dataframe with similarity-altered
counts, and offers the option to export results as .Rds or .csv files.
}
\examples{
results <- TCRsimilift_calculate(mouse_PBSvTCZ_data, sim_method="HAMMING", cutoff = 0.77, export_results=TRUE, output_directory="my_outputs")

}
