% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRsimilift_make_weighted_counts.R
\name{TCRsimilift_make_weighted_counts}
\alias{TCRsimilift_make_weighted_counts}
\title{Make count matrix from outputs of TCRsimilift_calculate}
\usage{
TCRsimilift_make_weighted_counts(new.data, doFilter = FALSE)
}
\arguments{
\item{new.data}{A dataframe outputted by \link{TCRsimilift_calculate}}

\item{doFilter}{Filter out sequences that do not have at least 2 counts? TRUE/FALSE.}
}
\value{
Count matrix containing weighted counts for each sequence and sample.
}
\description{
Make count matrix from outputs of TCRsimilift_calculate
}
