% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_counts.R
\name{weight_counts}
\alias{weight_counts}
\title{Add column of adjusted counts based on similarity}
\usage{
weight_counts(data.1, sim_method, cutoff = 0.8)
}
\arguments{
\item{data.1}{Dataframe with counts for all samples, AIRR format.}

\item{sim_method}{Either HAMMING or BLOSUM.}

\item{cutoff}{Minimum similarity score to consider two sequences neighbours. Between 0 and 1. Default 0.8 .}
}
\value{
data.1 with added column wrc of adjusted counts.
}
\description{
For each cdr3 amino acid length and V-J combination, a similarity matrix is
calculated once for all unique sequences in that group, then applied to each
sample. This avoids recomputing the similarity matrix per sample.
}
\details{
See \link{adjust_counts} for a full example of the DGE workflow.
}
