% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dist_matrix.R
\name{compute_dist_matrix}
\alias{compute_dist_matrix}
\title{Add column of adjusted counts based on similarity}
\usage{
compute_dist_matrix(data.1, z, sim_method)
}
\arguments{
\item{data.1}{Dataframe with counts, AIRR format.}

\item{z}{Sample size, ancillary.}

\item{sim_method}{Similarity method, either HAMMING or BLOSUM.}
}
\value{
data.1 with added column wrc of adjusted counts.
}
\description{
Add column of adjusted counts based on similarity
}
