% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck.R
\name{datacheck}
\alias{datacheck}
\title{Check if dataset has all necessary columns}
\usage{
datacheck(df)
}
\arguments{
\item{df}{Dataframe of immunological data.}
}
\description{
This internal function of the package verifies that the input dataframe contains the
needed AIRR columns. It does not verify their contents, only their existence.
}
\details{
The function will only have outputs if there are any errors or warnings, otherwise
the dataset is considered OK.

See \link{TCRsimilift_calculate} for a full example of the DGE workflow.
}
