% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{dataprep}
\alias{dataprep}
\title{Data preparation for simweightR}
\usage{
dataprep(df)
}
\arguments{
\item{df}{AIRR formatted dataframe of immunological data.}
}
\value{
A dataframe
}
\description{
Preprocessing the data for downstream use involves:
\itemize{
\item adds a column for cdr3 amino acid sequence lengths
\item aggregates all counts for each sequence id and each sample.
\item replaces NAs with 0.
\item returns aggregate data with only necessary associated columns.
}
}
\details{
See \link{adjust_counts} for a full example of the DGE workflow.
}
