% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_outputs.R
\name{export_outputs}
\alias{export_outputs}
\title{Export results of TCRsimilift}
\usage{
export_outputs(new.data, output_directory = "outputs")
}
\arguments{
\item{new.data}{Dataframe produced by the TCRsimilift_calculate function.}

\item{output_directory}{Name of output folder. "outputs" by default.}
}
\description{
Export results of TCRsimilift
}
