% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_update_data.R
\name{net_update_data}
\alias{net_update_data}
\title{Call compute_dist_matrix for each sample.}
\usage{
net_update_data(data, sim_method = "HAMMING")
}
\arguments{
\item{data}{Dataframe of AIRR format immunological data.}

\item{sim_method}{HAMMING or BLOSUM.}
}
\value{
Returns dataframe with adjusted counts.
}
\description{
Call compute_dist_matrix for each sample.
}
